+++ src/class/text/mod.rs
@@ -5,9 +5,10 @@ pub mod font;
 pub mod font_size;
 pub mod text_align;
 pub mod text_color;
+pub mod text_value;
 
 pub(super) mod text_prelude {
-    pub use super::{font::*, font_size::*, text_align::*, text_color::*};
+    pub use super::{font::*, font_size::*, text_align::*, text_color::*, text_value::*};
     pub use bevy_text::Text;
 }
 use text_prelude::*;
@@ -16,5 +17,6 @@ impl_class!(TextClass -> Text {
     FontHandle(FontHandle),
     FontSize(FontSize),
     JustifyText(JustifyText),
-    TextColor(TextColor)
+    TextColor(TextColor),
+    TextValue(TextValue)
 });